Fix gpcheckcat extraneous entry repair to use primary keys if oids not consistent across segments.
gpcheckcat missing/extraneous repair assumed oids are consistent across segments for a given catalog entry.
This is not always guaranteed for all catalog tables (which are defined in gpcheckcat). Instead, we need
to generate the repair based on the primary keys (the unique identifier in this case).

Authors: Marbin, Larry, Chris and Chumki